question_reply

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -27,7 +27,7 @@ public class PublishController {
         return "publish";
     }
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable Integer id,
+    public String edit(@PathVariable Long id,
                        Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("title",questionDTO.getTitle());
@@ -40,7 +40,7 @@ public class PublishController {
     public String doPublish(@RequestParam("title") String title,
                             @RequestParam("description") String description,
                             @RequestParam("tag") String tag,
-                            @RequestParam(name = "id",required = false) Integer id,
+                            @RequestParam(name = "id",required = false) Long id,
                             HttpServletRequest request,
                             Model model){
         model.addAttribute("title",title);
